refactor(svg): inline Q/E camera rAF loop in SvgGameView (refs #75)
Made-with: Cursor
